Added .gbdinit, cursor support to terminal
Updated the Makefile by adding itself to the dependencies of the
objects. This is for when the Makefile is updated, that all the objects
are re-compiled.

.gdbinit automatically tried to connect to a localhost:1234 for
debugging in qemu. It adds a few breakpoints too.

tty.c is updated to move the cursor when text is written to the screen.
This is useful for seeing where the current line is. When a newline is
reached, that line is also cleared so you don't have extra data on the
end of the line.